tests: Adding tests on plan
These tests should create a plan, apply it, and verify that the remote plan has been deleted
To normalize name, I changed RunPlan to PlanRun to have the json key : "plan_run". This will allow to have in the future a "plan_prefix" and keep the same prefix name for everything concerning plan
It also nsure that the plan file is deleted on the destroy action. This will be a rare usecase as the plan file will be deleted on each apply