destination directory for a terraform override file
This implements a new option `module_override_files` to specify a list of `src`and `dst` pairs, to copy a override file to a specific directory. e.g. to override a module

Following tests are implemented:
* overrides the existing resource definition in module
* errors when given a directory for source
* errors when given a file for destination
* errors when given an invalid path for source
* errors when given an invalid path for destination
* errors when src key is missing
* errors when dst key is missing

The destination is relative to the concourse working directory to accommodate for modules being in a parent directory.

example:
```
terraform_source: fixtures/module/
module_override_files:
- src: fixtures/override/example_override.tf
  dst: fixtures/aws
```