Add Terraform version as a label to Docker image
Having this value available in the image allows downstream consumers to
easily parse the value from the image when building new images.

To achieve this, the pre-existing `push-prod-image` job in the Concourse
pipeline has been extended. The `write-docker-tags` task has been
renamed appropriately and now also constructs a JSON file of labels to-
apply to the image being built.

It was chosen to do this in the Concourse pipeline as it is the one
opportunity during the build pipeline that the Terraform version is
availble and can be passed to the Dockerfile/`docker build` process as a
label.