Add curl to images
Curl is a super valuable tool to have available for debugging and
execution, as certain Terraform modules assume the tool is available on
the system. Wget is less common on systems and most tools assume the
existence of curl before wget.

For example: wait_for_cluster_command in CloudPosse's EKS cluster module
uses curl by default. While this can be overridden to use wget, not all
the environments we run Terraform in have wget installed and we would
prefer to not introduce a bunch of conditional logic:
https://github.com/cloudposse/terraform-aws-eks-cluster/blob/783799fa64f20cde64ae19e74e046263072c8595/variables.tf#L142